new: ExitManager for common logic related to building proof
