+++ src/common/ExitManager.ts
@@ -1,0 +1,98 @@
+import BN from 'bn.js'
+import assert from 'assert'
+import ethUtils from 'ethereumjs-util'
+
+import Web3Client from './Web3Client'
+import ContractsBase from './ContractsBase'
+import RootChain from '../root/RootChain'
+
+import Proofs from 'matic-protocol/contracts-core/helpers/proofs.js'
+
+const logger = {
+  info: require('debug')('maticjs:WithdrawManager'),
+  debug: require('debug')('maticjs:debug:WithdrawManager'),
+}
+
+export default class ExitManager extends ContractsBase {
+  private rootChain: RootChain
+
+  constructor(rootChain: RootChain, web3Client: Web3Client) {
+    super(web3Client)
+    this.rootChain = rootChain
+  }
+
+  async buildPayloadForExit(burnTxHash, logEventSig) {
+    // check checkpoint
+    const lastChildBlock = await this.rootChain.getLastChildBlock()
+    const burnTx = await this.web3Client.getMaticWeb3().eth.getTransaction(burnTxHash)
+    const receipt = await this.web3Client.getMaticWeb3().eth.getTransactionReceipt(burnTxHash)
+    const block: any = await this.web3Client
+      .getMaticWeb3()
+      .eth.getBlock(burnTx.blockNumber, true /* returnTransactionObjects */)
+
+    logger.info({ 'burnTx.blockNumber': burnTx.blockNumber, lastCheckPointedBlockNumber: lastChildBlock })
+    assert.ok(
+      new BN(lastChildBlock).gte(new BN(burnTx.blockNumber)),
+      'Burn transaction has not been checkpointed as yet'
+    )
+    const headerBlockNumber = await this.rootChain.findHeaderBlockNumber(burnTx.blockNumber)
+    const headerBlock = await this.web3Client.call(
+      this.rootChain.getRawContract().methods.headerBlocks(this.encode(headerBlockNumber))
+    )
+    logger.info({ headerBlockNumber: headerBlockNumber.toString(), headerBlock })
+
+    // build block proof
+    const blockProof = await Proofs.buildBlockProof(
+      this.web3Client.getMaticWeb3(),
+      headerBlock.start,
+      headerBlock.end,
+      burnTx.blockNumber
+    )
+
+    const receiptProof = await Proofs.getReceiptProof(receipt, block, this.web3Client.getMaticWeb3())
+
+    const logIndex = receipt.logs.findIndex(log => log.topics[0].toLowerCase() == logEventSig.toLowerCase())
+    assert.ok(logIndex > -1, 'Log not found in receipt')
+
+    return this._encodePayload(
+      headerBlockNumber,
+      blockProof,
+      burnTx.blockNumber,
+      block.timestamp,
+      Buffer.from(block.transactionsRoot.slice(2), 'hex'),
+      Buffer.from(block.receiptsRoot.slice(2), 'hex'),
+      Proofs.getReceiptBytes(receipt), // rlp encoded
+      receiptProof.parentNodes,
+      receiptProof.path,
+      logIndex
+    )
+  }
+
+  private _encodePayload(
+    headerNumber,
+    buildBlockProof,
+    blockNumber,
+    timestamp,
+    transactionsRoot,
+    receiptsRoot,
+    receipt,
+    receiptParentNodes,
+    path,
+    logIndex
+  ) {
+    return ethUtils.bufferToHex(
+      ethUtils.rlp.encode([
+        headerNumber,
+        buildBlockProof,
+        blockNumber,
+        timestamp,
+        ethUtils.bufferToHex(transactionsRoot),
+        ethUtils.bufferToHex(receiptsRoot),
+        ethUtils.bufferToHex(receipt),
+        ethUtils.bufferToHex(ethUtils.rlp.encode(receiptParentNodes)),
+        ethUtils.bufferToHex(ethUtils.rlp.encode(path)),
+        logIndex,
+      ])
+    )
+  }
+}
